+++ src/App.tsx
@@ -1,14 +1,18 @@
-import React from 'react';
-
 import './styles.css';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
 import { Landing } from '@homework-task/components/landing/Landing';
 
-const App: React.FC = () => {
+const queryClient = new QueryClient();
+
+function App() {
     return (
         <main>
-            <Landing />
+            <QueryClientProvider client={queryClient}>
+                <Landing />
+            </QueryClientProvider>
         </main>
     );
-};
+}
 
 export default App;