+++ src/components/landing/Landing.tsx
@@ -2,38 +2,9 @@ import React from 'react';
 
 import clsx from 'clsx';
 
-import { Card } from './Card';
-
-interface CardData {
-    title: string;
-    text: string;
-    link: string;
-}
-
-const vscodeProjectLink: string = 'vscode://file/${__CWD__}'; // Ensure the template string is used correctly in the actual code.
-
-const cards: CardData[] = [
-    {
-        title: 'Transfer the project to TypeScript',
-        text: 'Your first task involves transitioning this project from JavaScript to TypeScript.',
-        link: vscodeProjectLink,
-    },
-    {
-        title: 'Create a scalable List Component',
-        text: 'Build a scalable React component to fetch and display key information (id, name, username, email, and phone) from an API in a list format.',
-        link: vscodeProjectLink,
-    },
-    {
-        title: 'Create a Form Generator Component',
-        text: 'Build a versatile React component with validation, API hook, and form rendering capabilities.',
-        link: vscodeProjectLink,
-    },
-    {
-        title: 'Create a Page Generator Component',
-        text: 'Create a versatile React component for dynamic webpage construction, adapting to various layouts and components through received props.',
-        link: vscodeProjectLink,
-    },
-];
+import getUsers from '@homework-task/api/users';
+import List from '@homework-task/components/List';
+import ListCard from '@homework-task/components/ListCard';
 
 export const Landing: React.FC = () => {
     return (
@@ -47,7 +18,8 @@ export const Landing: React.FC = () => {
                 'gap-16',
                 'place-content-center',
                 'p-8',
-                'sm:p-16'
+                'sm:p-16',
+                'relative'
             )}
         >
             <div
@@ -60,6 +32,17 @@ export const Landing: React.FC = () => {
                     'w-full'
                 )}
             >
+                <img
+                    className={clsx(
+                        'w-[148px]',
+                        'h-[168px]',
+                        'absolute',
+                        'left-[-104px]',
+                        'rotate-45'
+                    )}
+                    src="/media/cats/cat_10.png"
+                    alt="Cat"
+                />
                 <h1
                     className={clsx(
                         'leading-normal',
@@ -69,31 +52,12 @@ export const Landing: React.FC = () => {
                         'text-black'
                     )}
                 >
-                    Welcome to Fat Cats{' '}
+                    Welcome to{' '}
                     <span className={clsx('text-primary')}>
-                        Homework Task Page
+                        Radojica Popovic&apos;s fatcat-task
                     </span>
-                    -where curiosity meets opportunity.
+                    - where curiosity meets opportunity.
                 </h1>
-                <p className={clsx('text-gray80', 'text-2xl')}>
-                    Dont hesitate to connect if you have any questions.
-                </p>
-                <a
-                    className={clsx(
-                        'flex',
-                        'items-center',
-                        'gap-2',
-                        'rounded-lg',
-                        'px-4',
-                        'py-2',
-                        'bg-black',
-                        'text-white'
-                    )}
-                    href={vscodeProjectLink}
-                >
-                    <span className={clsx('text-lg')}>Read docs</span>
-                    <img src="/media/landing/arrow.svg" alt="" />
-                </a>
             </div>
             <img
                 className={clsx(
@@ -104,8 +68,9 @@ export const Landing: React.FC = () => {
                     'justify-self-center'
                 )}
                 src="/media/landing/hero.svg"
-                alt=""
+                alt="Hero image"
             />
+
             <div
                 className={clsx(
                     'col-span-full',
@@ -115,11 +80,13 @@ export const Landing: React.FC = () => {
                     'md:grid-cols-2',
                     'xl:grid-cols-4'
                 )}
-            >
-                {cards.map((card) => (
-                    <Card key={card.title} {...card} />
-                ))}
-            </div>
+            ></div>
+            <List
+                queryKey={['users']}
+                queryFn={getUsers}
+                ItemComponent={ListCard}
+                errorMessage="Failed to fetch users."
+            />
         </section>
     );
 };