+++ src/types/user.type.ts
@@ -1,0 +1,9 @@
+type User = {
+    id: number;
+    name: string;
+    username: string;
+    email: string;
+    phone: string;
+};
+
+export default User;