+++ CHANGELOG.md
@@ -5,12 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 02-07-2024
+## [1.0.2] - 02-07-2024
+### Added
+- Created a List Component
+    - added react-query to the App.tsx
+    - fetching the data from api and displaying it
+    - added one extra cat to Landing page | changed layout of the page
+
+## [1.0.1] - 02-07-2024
 ### Added
 - Conversion from JavaScript to Typescript | Changes to Project Structure
     - added tsconsig.json
     - edited configuration
     - installed dependencies
+    
 
 
 ## [1.0.0] - 02-07-2024